// demos/convex/node/convex/_generated/api.ts
// Generated by @nimbus/codegen. Do not edit by hand.
import { makeMutationReference, makeQueryReference } from "convex/browser";
import type { Doc, Id } from "./dataModel";

type JsonValue = null | boolean | number | string | JsonValue[] | { [key: string]: JsonValue };

export const api = {
  messages: {
    list: makeQueryReference<{}, Doc<"messages">[]>("messages:list", "public"),
    send: makeMutationReference<{
  "author": string;
  "body": string;
}, Id<"messages">>("messages:send", "public")
  }
};

export const internal = {};